fixed modal window bug
modal window was giving "could not read properties of undefined, reading props.id". This was because when creating a react app with a constructor, props should be used to set the state, then the state is used in render. props is not in the scope of render. Fixed by removing the unnecessary constructor.